requirements.txt
fastapi[standard]
pymongo
openai
google-generativeai
python-dotenv
chromadb
googlesearch-python